This commit was manufactured by cvs2svn to create tag 'TAG_DNPROGS_2_20'.

+++ dnprogs/dndir/dndir.cc
@@ -616,7 +616,7 @@ void print_short(
 	    }
 	    else
 	    {
-		printf("%-*s",(int)(filename_width-(strlen(name)-filename_width)), "");
+		printf("%-*s",filename_width-(strlen(name)-filename_width), "");
 	    }
 	    ++(*printed);
 	}
